+++ app/feeds/show/template.hbs
@@ -69,9 +69,9 @@
   {{#each model.issues as |issue|}}
     {{#collapsable-panel title="Error fetching feed..." panelClass="panel-danger"}}
       <strong>Issue created by feed fetch {{moment-from-now model.feed.last_fetched_at}} ago:</strong>
-      <pre>{{#link-to 'issues.feed-fetch.show' issue.id}}
+      <!-- <pre>{{#link-to 'issues.feed-fetch.show' issue.id}}
         Feed Fetch Issue: #{{issue.id}}
-      {{/link-to}}</pre>
+      {{/link-to}}</pre> -->
       {{#edit-buttons showChangeset="showChangeset" closeDialog="closeDialog"}}{{/edit-buttons}}
     {{/collapsable-panel}}
   {{/each}}
