+++ app/issues/index/template.hbs
@@ -18,9 +18,6 @@
       <div class="col-md-6 issues-link">
         {{#link-to "issues.route-geometry" (query-params feed_onestop_id=selectedFeed) class="list-group-item active"}}Route Geometry{{/link-to}}
       </div>
-      <div class="col-md-6 issues-link">
-        {{#link-to "issues.feed-fetch" class="list-group-item active"}}Feed Fetch Issues{{/link-to}}
-      </div>
     </div>
   </div>
 </div>
